Pictures of Products (2 dimensions)
Placed into the grid full page, I think we need to add 3 more for aesthetics but let's do that at the end if we have more time. If Ahmed hasn't created the product pages I'll start doing some of that next.

Note that the images for the product pages are there in the correct dimensions